// src/common/enums/parameter-map-category.enum.ts
// src/common/enums/parameter-map-category.enum.ts
export enum ParameterMapCategory {
  text_to_text = 'text-to-text',
  image_to_image = 'image-to-image',
  video_to_video = 'video-to-video',
  audio_to_audio = 'audio-to-audio',
  text_to_image = 'text-to-image',
  text_to_video = 'text-to-video',
  text_to_audio = 'text-to-audio',
  image_to_text = 'image-to-text',
  image_to_video = 'image-to-video',
  image_to_audio = 'image-to-audio',
  video_to_text = 'video-to-text',
  video_to_image = 'video-to-image',
  video_to_audio = 'video-to-audio',
  audio_to_text = 'audio-to-text',
  audio_to_image = 'audio-to-image',
  audio_to_video = 'audio-to-video',
}